% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciate.R
\name{find_code}
\alias{find_code}
\title{Find PROFILE_CODE}
\usage{
find_code(profile, by = "Keywords")
}
\arguments{
\item{profile}{Character, to search PROFILE CODE}

\item{by}{Character, to search code. eg: "Keywords", "PROFILE_NOTES", "PROFILE_TYPE"
or other name of PROFILES}
}
\value{
a data.frame with with profile codes
}
\description{
\code{\link{find_code}} Return a data.frame with profile codes
}
\examples{
\dontrun{
profile <- "Ethanol"
dt <- find_code(profile)
}
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
