% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respeciate-package.R
\docType{package}
\name{respeciate-package}
\alias{respeciate}
\alias{respeciate-package}
\title{respeciate: Speciation profiles for gases and aerosols}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Acess to the US.EPA Speciate (v5.1) tool,
    to generate speciation for gases and particles. 
    More details in Simon et al (2010) <doi:10.5094/APR.2010.026>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atmoschem/respeciate}
  \item Report bugs at \url{https://github.com/atmoschem/respeciate/issues}
}

}
\author{
\strong{Maintainer}: Sergio Ibarra-Espinosa \email{sergio.ibarra@usp.br} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})

}
\keyword{internal}
