% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciate.0.2.R
\name{respeciate}
\alias{respeciate}
\alias{find_profile_code}
\alias{profile}
\alias{get_profile}
\alias{print.respeciate}
\alias{plot.respeciate}
\alias{print.respeciate.ref}
\title{respeciate}
\usage{
find_profile_code(term, by = "Keywords")

get_profile(code)

\method{print}{respeciate}(x, ...)

\method{plot}{respeciate}(x, ...)

\method{print}{respeciate.ref}(x, ...)
}
\arguments{
\item{term}{character, the search term to use
when searching archive.}

\item{by}{character, the section of the archive to
search, by default \code{'keywords'}.}

\item{code}{character or numeric, the speciate code
of the required source profile (EPA SPECIATE PROFILE_CODE).}

\item{x}{the \code{respeciate.ref} object to be printed.}

\item{...}{any extra arguments, currently ignored.}
}
\value{
\code{find_profile_code} returns a object of
\code{respeciate.ref} class, a \code{data.frame} of
profile information.

\code{get_profile} returns a object of
\code{respeciate} class, a \code{data.frame} containing a
speciate profile.
}
\description{
respeciate but with object classes...

\code{\link{find_profile_code}} finds
the codes of speciate profiles in the local speciate
archive using supplied search terms.

\code{\link{get_profile}} extracts a
speciate profile from the local speciate archive.

When supplied a \code{respeciate}
object, \code{\link{print}} manages its appearance.

When supplied a \code{respeciate}
object, \code{\link{plot}} provides a basic plot
output. The objects is a data.frame so you can still
use it with as normal with \code{lattice} or
\code{ggplot2}...

When supplied a \code{respeciate.ref}
object, \code{\link{print}} manages its appearance.
}
\examples{
\dontrun{
profile <- "Ethanol"
dt <- find_profile_code(profile)
}
\dontrun{
code <- "8855"
x <- spec(code)
}
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
