% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.combine.R
\name{rsp.combine}
\alias{rsp.combine}
\alias{rsp_lbind}
\title{combining respeciate profiles}
\usage{
rsp_lbind(...)
}
\arguments{
\item{...}{(various) This function is intended to be quite flexible. All
supplied arguments are tested and handled as follows: \code{respeciate}-like
objects are passed to \code{data.table::rbindlist} as a list to rbind using
\code{data.table} methods; Any other arguments that are valid
\code{rbindlist} arguments are passed on 'as is'; And, anything else is
(hopefully) ignored.}
}
\value{
\code{rsp_lbind} attempts to return a single stacked version of the
supplied data sets. If it is successful, the (stacked) data set is typically
returned as a \code{respeciate} object or a \code{data.frame} with a warning
if it is missing columns \code{respeciate} expects.
}
\description{
Functions to combining respeciate data sets.

\code{rsp_lbind} binds two or more \code{respeciate}-like
objects. The default option is to stack the supplied data sets (e.g.
\code{respeciate}, \code{data.frame}, etc) like \code{rbindlist}
in \code{data.table} (or row_bind in dplyr). This matches columns by name
before stacking the supplied data sets.
}
\references{
Dowle M, Srinivasan A (2023). data.table: Extension of `data.frame`.
  R package version 1.14.8, \url{https://CRAN.R-project.org/package=data.table}.
}
