% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.eu.R
\name{rsp.eu}
\alias{rsp.eu}
\alias{rsp_eu}
\alias{rsp_eu_pm10}
\alias{rsp_eu_pm2.5}
\title{Quick access to common SPECIAEUROPE subsets.}
\usage{
rsp_eu()

rsp_eu_pm10()

rsp_eu_pm2.5()
}
\value{
\code{rsp_eu} and \code{rsp_eu_}functions typically return
a \code{respeciate} \code{data.frame} of the requested profiles:

\code{rsp_eu()} returns all profiles in the local version of
\code{\link{SPECIEUROPE}}

\code{rsp_eu_pm10} returns all SPECIEUROPE profiles classified as
PM10 (using \code{Particle.Size=="PM10"}), \code{rsp_eu_pm10} for PM2.5
and so on...
}
\description{
\code{\link{rsp_eu}} and \code{rsp_eu_} functions are
quick access wrappers to commonly requested SPECIEUROPE subsets.
}
\seealso{
\code{\link{SPECIEUROPE}}
}
