% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.export.R
\name{rsp.export}
\alias{rsp.export}
\alias{rsp_export_esat}
\title{Exporting respeciate objects}
\usage{
rsp_export_esat(
  rsp,
  file.name = "file",
  index = "row.count",
  unc = 0.15,
  bad.values = "fill.1",
  output = c("con.csv", "unc.csv"),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{rsp}{(\code{respeciate} or similar, e.g. a data-frame set up for use
with \code{respeciate}), the data-set to export.}

\item{file.name}{(\code{character}), the file name of the exported file or
files. See also \code{output} and Details below.}

\item{index}{(\code{character}), the name of \code{rsp} column to use as the
output file(s) index or (default) \code{'row.count'} a row number counter.}

\item{unc}{(various), if \code{numeric}, the scaling factor to apply to
concentration values when hole filling uncertainties, else \code{'eu.rsp'},
in which case it tries to recover values from the \code{SPECIEUROPE} meta
information.}

\item{bad.values}{(\code{character}), handling method to use if bad values
are found in the supplied data.}

\item{output}{(\code{character}), the file types to export. See also
Details below.}

\item{overwrite}{(\code{character}), overwrite file if it already exists.}

\item{...}{other arguments, currently ignored.}
}
\value{
\code{rsp_export}s attempt to build and save files suitable for use
outside \code{r}.
}
\description{
rsp function(s) to export respeciate (and respeciate-like)
objects to other software
}
\details{
\code{rsp_build_esat} makes files that can be used as inputs with ESAT.
\code{output} options: 'con.csv' and 'unc.csv' (both required by ESAT).
}
