% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.id.R
\name{rsp.id}
\alias{rsp.id}
\alias{rsp_id_copy}
\alias{rsp_id}
\alias{rsp_id_nalkane}
\alias{rsp_id_btex}
\alias{rsp_id_pah16}
\title{rsp_id_ functions to identify common species groups for grouping and
subsetting respeciate profiles}
\usage{
rsp_id_copy(rsp, ref = NULL, by = ".species.id")

rsp_id_nalkane(rsp)

rsp_id_btex(rsp)

rsp_id_pah16(rsp)
}
\arguments{
\item{rsp}{a \code{respeciate} object, a \code{data.frame} of respeciate
profiles.}

\item{ref}{(\code{rsp_id_copy} only) a second \code{respeciate} object, to
be used as reference when subsetting (or conditioning) \code{rsp}.}

\item{by}{(\code{rsp_id_copy} only) character, the name of the column
in \code{ref} to copy when subsetting (or conditioning) \code{rsp}.}
}
\value{
\code{rsp_id_copy} outputs can be modified but, by default, it
identifies all species in the supplied reference data set.

\code{rsp_id_nalkane} identifies (straight chain) C1 to C40 n-alkanes.

\code{rsp_id_btex} identifies the BTEX group of aromatic hydrocarbons
(benzene, toluene, ethyl benzene, and M-, O- and P-xylene).
}
\description{
\code{rsp_id_} functions generate a vector of assignment
terms and can be used to subset or condition a supplied (re)SPECIATE
\code{data.frame}.

Most commonly, the \code{rsp_id_} functions accept a single input, a
respeciate \code{data.frame} and return a logical vector of
length \code{nrow(x)}, identifying species of interest as
\code{TRUE}. So, for example, they can be used when
\code{\link{subset}}ting in the form:

\code{subset(rsp, rsp_id_nalkane(rsp))}

... to extract just n-alkane records from a supplied \code{respeciate}
object \code{rsp}.

However, some accept additional arguments. For example, \code{rsp_id_copy}
also accepts a reference data set, \code{ref}, and a column identifier,
\code{by}, and tests \code{rsp$by \%in\% unique(ref$by)}.
}
