% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.info.R
\name{rsp.info}
\alias{rsp.info}
\alias{rsp_info}
\title{Information about data sets currently in respeciate}
\usage{
rsp_info()
}
\value{
\code{rsp_info} provides a brief version information report on the
currently installed respeciate archive.
}
\description{
Functions that provide respeciate
source information.
\code{rsp_info} generates a brief version report for the currently installed
respeciate data sets.
}
\examples{
\dontrun{
rsp_info()
}

}
\references{
For SPECIATE:

Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.

For SPECIEUROPE:

Pernigotti, D., Belis, C.A., Spano, L., 2016. SPECIEUROPE: The
European data base for PM source profiles. Atmospheric Pollution Research,
7(2), pp.307-314. DOI: https://doi.org/10.1016/j.apr.2015.10.007
}
\seealso{
\code{\link{SPECIATE}} and \code{\link{SPECIEUROPE}}
}
