% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.plot.R
\name{rsp.plot}
\alias{rsp.plot}
\alias{rsp_plot_profile}
\alias{rsp_plot}
\alias{rsp_plot_species}
\title{plotting respeciate source profiles}
\usage{
rsp_plot_profile(
  rsp,
  id,
  multi.profile = "group",
  order = TRUE,
  log = FALSE,
  ...,
  silent = FALSE,
  output = "default"
)

rsp_plot_species(
  rsp,
  id,
  multi.species = "group",
  order = FALSE,
  log = FALSE,
  ...,
  silent = FALSE,
  output = "default"
)
}
\arguments{
\item{rsp}{A \code{respeciate} object, a \code{data.frame} of respeciate
profiles.}

\item{id}{numeric, the indices of profiles or species to use when
plotting with \code{rsp_plot_profile} or \code{rsp_plot_species},
respectively. For example, \code{rsp_plot_profile(rsp, id=1:6)} plots
first 6 profiles in \code{respeciate} object \code{rsp}.}

\item{multi.profile}{character, how \code{rsp_plot_profile} should
handle multiple profiles, e.g. 'group' or 'panel' (default
group).}

\item{order}{logical, order the species in the
profile(s) by relative abundance before plotting.}

\item{log}{logical, log y scale when plotting.}

\item{...}{any additional arguments, typically passed on the lattice
plotting functions.}

\item{silent}{logical, hide warnings when generating plots (default
\code{FALSE})}

\item{output}{character, output method, one of: 'plot' to return just the
requested plot; 'data' to return just the data; and, c('plot', 'data') to
plot then return the data invisibly (default).}

\item{multi.species, }{character, like \code{multi.profile} in
\code{sp_plot_profile} but for species in \code{sp_plot_species}.}
}
\value{
\code{sp_plot} graph, plot, etc usually as a trellis object.
}
\description{
General plots for \code{respeciate} objects.

\code{rsp_plot} functions generate plots for supplied
respeciate data sets.
}
\note{
These functions are currently in development, so may change.
}
\references{
Most \code{respeciate} plots make extensive use of
\code{lattice} and \code{latticeExtra} code:

Sarkar D (2008). \emph{Lattice: Multivariate Data Visualization with R}.
Springer, New York. ISBN 978-0-387-75968-5, \url{http://lmdvr.r-forge.r-project.org}.

Sarkar D, Andrews F (2022). \emph{latticeExtra: Extra Graphical Utilities Based
on Lattice}. R package version 0.6-30,
\url{https://CRAN.R-project.org/package=latticeExtra}.

They also incorporate ideas from \code{loa}:

Ropkins K (2023). \emph{loa: various plots, options and add-ins for use with lattice}.
R package version 0.2.48.3, \url{https://CRAN.R-project.org/package=loa}.
}
