% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.pls.plot.R
\name{rsp.pls.plot}
\alias{rsp.pls.plot}
\alias{pls_plot}
\alias{pls.plot}
\alias{pls_plot_species}
\alias{pls_plot_profile}
\title{Plots for use with (re)SPECIATE profile Positive Least Squares models}
\usage{
pls_plot(pls, plot.type = 1, ..., output = "default")

pls_plot_profile(pls, plot.type = 1, log = FALSE, ..., output = "default")

pls_plot_species(pls, id, plot.type = 1, ..., output = "default")
}
\arguments{
\item{pls}{A \code{sp_pls_profile} output, intended for use with
\code{pls_} functions.}

\item{plot.type}{numeric, the plot type if
multiple options are available.}

\item{...}{other arguments, typically passed on to the associated
\code{lattice} plot.}

\item{output}{character, output method, one of: 'plot' to return just the
requested plot; 'data' to return just the data; and, c('plot', 'data') to
plot then return the data invisibly (default).}

\item{log}{(for \code{pls_plot_profile} only) logical, if \code{TRUE} this
applies 'log' scaling to the primary Y axes of the plot.}

\item{id}{numeric or character
identifying the species or profile to plot. If numeric, these are treated
as indices of the species or profile, respectively, in the PLS model; if
character, species is treated as the name of species and profile is treated
as the profile code. Both can be concatenated to produce multiple plots and
the special case \code{id = -1} is a short cut to all species or profiles,
respectively.}
}
\value{
\code{pls_plot}s produce various plots commonly used in source
apportionment studies.
}
\description{
The \code{pls_plot} functions are intended for use with PLS models built
using \code{rsp_pls_profile} (documented separately). They generate some
plots commonly used with source apportionment model outputs.
}
