% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.reshape.R
\name{rsp.reshape}
\alias{rsp.reshape}
\alias{rsp_dcast}
\alias{rsp_dcast_profile}
\alias{rsp_dcast_species}
\alias{rsp_melt_wide}
\title{Reshaping respeciate data sets}
\usage{
rsp_dcast(rsp, widen = "species")

rsp_dcast_profile(rsp, widen = "profile")

rsp_dcast_species(rsp = rsp, widen = "species")

rsp_melt_wide(rsp, pad = TRUE, drop.nas = TRUE)
}
\arguments{
\item{rsp}{A \code{respeciate} object, a \code{data.frame} of respeciate
profiles in standard long form or widened form using
\code{\link{rsp_dcast}} and \code{\link{rsp_melt_wide}}, respectively.}

\item{widen}{character, when widening \code{rsp} with
\code{\link{rsp_dcast}}, the data type to \code{dcast},
currently \code{'species'} (default) or \code{'profile'}. See Note.}

\item{pad}{logical or character, when \code{melt}ing a previously widened
data set, should output be re-populated with species and/or profile
meta-data, discarded when widening. This is currently handled by
\code{\link{rsp_pad}}. The default \code{TRUE} applies standard settings,
so does not include profile sources reference meta-data. (See
\code{\link{rsp_pad}} for other options).}

\item{drop.nas}{logical, when \code{melt}ing a previously widened
data set, should output be stripped of any rows containing empty
weight/value columns. Because not all profile contains all species, the
\code{dcast}/\code{melt} process can generate empty rows, and this step
attempt account for that when working with standard re(SPECIATE)
profiles. It is, however, sometimes useful to check first, e.g. when
building profiles yourself.}
}
\value{
\code{rsp_dcast} returns the wide form of the supplied
\code{respeciate} profile. \code{rsp_melt_wide}
returns the (standard) long form of a previously widened profile.
}
\description{
Functions for reshaping respeciate profiles

\code{rsp_dcast} and \code{rsp_melt_wide} reshape supplied
respeciate profile(s). \code{rsp_dcast} converts these from their supplied
long form to a widened form, \code{dcast}ing the data set by either species
or profiles depending on the \code{widen} setting applied.
\code{rsp_dcast_profile} and \code{rsp_dcast_species} are wrappers for these
options. \code{rsp_melt_wide} attempts to return a previously widened data
set to the original long form.
}
\note{
Conventional long-to-wide reshaping of data, or \code{dcast}ing, can
be slow and memory inefficient. So, \code{respeciate} uses the
\code{\link[data.table:dcast]{data.table::dcast}}
method. The \code{rsp_dcast_species} method,
applied using \code{widen='species'}, is effectively:

\code{dcast(..., PROFILE_CODE+PROFILE_NAME~SPECIES_NAME, value.var="WEIGHT_PERCENT")}

And, the alternative \code{widen='profile'}:

\code{dcast(..., SPECIES_ID+SPECIES_NAME~PROFILE_CODE, value.var="WEIGHT_PERCENT")}

Although, \code{respeciate} uses a local version of \code{WEIGHT_PERCENT} called
\code{.value}, so the EPA source information can easily be recovered. See also
\code{\link{rsp_rescale_profile}}.
}
\references{
Dowle M, Srinivasan A (2023). _data.table: Extension of `data.frame`_.
  R package version 1.14.8, <https://CRAN.R-project.org/package=data.table>.
}
