% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.us.R
\name{rsp.us}
\alias{rsp.us}
\alias{rsp_us_gas}
\alias{rsp_us}
\alias{rsp_us_other}
\alias{rsp_us_pm}
\alias{rsp_us_pm.ae6}
\alias{rsp_us_pm.ae8}
\alias{rsp_us_pm.cr1}
\alias{rsp_us_pm.simplified}
\title{Quick access to common SPECIATE subsets.}
\usage{
rsp_us_gas()

rsp_us_other()

rsp_us_pm()

rsp_us_pm.ae6()

rsp_us_pm.ae8()

rsp_us_pm.cr1()

rsp_us_pm.simplified()
}
\value{
\code{rsp_us_} functions typically return a \code{respeciate}
\code{data.frame} of the requested profiles.

For example:

\code{rsp_us_gas()} returns all gaseous profiles in SPECIATE
(\code{PROFILE_TYPE == 'GAS'}).

\code{rsp_us_pm} returns all particulate matter (PM) profiles in SPECIATE
not classified as a special PM type (\code{PROFILE_TYPE == 'PM'}).

The special PM types are subsets profiles intended for special
applications, and these include \code{rsp_us_pm.ae6} (type \code{PM-AE6}),
\code{rsp_us_pm.ae8} (type \code{PM-AE8}), \code{rsp_us_pm.cr1} (type
\code{PM-CR1}), and \code{rsp_us_pm.simplified} (type \code{PM-Simplified}).

\code{rsp_us_other} returns all profiles classified as other in SPECIATE
(\code{PROFILE_TYPE == 'OTHER'}).
}
\description{
\code{rsp_us_} functions are quick access wrappers to commonly
requested SPECIATE subsets.
}
