% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.x.R
\name{rsp.x}
\alias{rsp.x}
\alias{rsp_x_copy}
\alias{rsp_x}
\alias{rsp_x_nalkane}
\alias{rsp_x_btex}
\title{rsp_x_ functions for grouping and subsetting (re)SPECIATE profiles}
\usage{
rsp_x_copy(rsp, ref = NULL, by = "species_id")

rsp_x_nalkane(rsp)

rsp_x_btex(rsp)
}
\arguments{
\item{rsp}{a \code{respeciate} object, a \code{data.frame} of (re)SPECIATE
profiles.}

\item{ref}{(\code{rsp_x_copy} only) a second \code{respeciate} object, to
be used as reference when subsetting (or conditioning) \code{rsp}.}

\item{by}{(\code{rsp_x_copy} only) character, the name of the column
in \code{ref} to copy when subsetting (or conditioning) \code{rsp}.}
}
\value{
\code{rsp_x_copy} outputs can be modified but, by default, it
identifies all species in the supplied reference data set.

\code{rsp_x_nalkane} identifies (straight chain) C1 to C40 n-alkanes.

\code{rsp_x_btex} identifies the BTEX group of aromatic hydrocarbons
(benzene, toluene, ethyl benzene, and M-, O- and P-xylene).
}
\description{
\code{rsp_x_} functions generate a vector of assignment
terms and can be used to subset or condition a supplied (re)SPECIATE
\code{data.frame}.

Most commonly, the \code{rsp_x_} functions accept a single input, a
(re)SPECIATE \code{data.frame} and return a logical vector of
length \code{nrow(x)}, identifying species of interest as
\code{TRUE}. So, for example, they can be used when
\code{\link{subset}}ting in the form:

\code{subset(rsp, rsp_x_nalkane(rsp))}

... to extract just n-alkane records from a supplied \code{respeciate}
object \code{rsp}.

However, some accept additional arguments. For example, \code{rsp_x_copy}
also accepts a reference data set, \code{ref}, and a column identifier,
\code{by}, and tests \code{rsp$by \%in\% unique(ref$by)}.
}
