% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.average.R
\name{sp.average}
\alias{sp.average}
\alias{sp_average_profile}
\title{speciate data averaging functions}
\usage{
sp_average_profile(x, code = NULL, name = NULL, method = 1, ...)
}
\arguments{
\item{x}{A \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}

\item{code}{required character, the unique profile code to assign to the
average profile.}

\item{name}{character, the profile name to assign to the average
profile. If not supplied, this defaults to a collapsed list of the codes
of all the profiles averaged.}

\item{method}{numeric, the averaging method to apply: Currently only 1 (default)
\code{mean(x)}.}

\item{...}{additional arguments, currently ignored}
}
\value{
\code{sp_average_profile} returns a single profile average
version of the supplied \code{respeciate} profile.
}
\description{
Functions to build composite (re)SPECIATE profiles

\code{sp_average_profile} generates an average composite
of a supplied multi-profile \code{respeciate} object.
}
\note{
In development function; arguments and outputs likely to be subject to
change.

This is one of the very few \code{respeciate} functions that modifies the
\code{WEIGHT_PERCENT} column of the \code{respectiate} \code{data.frame}.
}
