% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.build.R
\name{sp.build}
\alias{sp.build}
\alias{sp_build_rsp_x}
\title{sp_build functions}
\usage{
sp_build_rsp_x(
  x,
  profile_code,
  profile_name,
  species_name,
  species_id,
  value,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} or similar (i.e.
something that can be coerced into a \code{data.frame} using
\code{as.data.frame}) to be converted into a \code{respeciate} object.}

\item{profile_name, profile_code}{(\code{character}) The name of the column
in \code{x} containing profile name and code records, respectively. If not
already named according to SPECIATE conventions, at least one of these will
need to be assigned.}

\item{species_name, species_id}{(\code{character}) The name of the column
in \code{x} containing species name and id records, respectively. If not
already named according to SPECIATE conventions, at least one of these will
need to be assigned.}

\item{value}{(\code{character})  The name of the column in \code{x}
containing measurement values. If not already named according to SPECIATE
conventions, this will need to be assigned.}
}
\value{
\code{sp_build}s attempt to build and return a (re)SPECIATE-like
profile that can be compared with data in re(SPECIATE).
}
\description{
sp function(s) to reconfigure data.frames (and similar
object classes) for use with data and functions in re(SPECIATE).
}
\note{
If you want to compare your data with profiles in the SPECIATE archive,
you need to use EPA SPECIATE conventions when assigning species names and
identifiers. Currently, we are working on options to improve on this (and
very happy to discuss if anyone has ideas), but current best suggestion is:
(1) identify the SPECIATE species code for each of the species in your data set,
and (2) assign these as \code{species_id} when \code{sp_build}ing. The
function will then associate the \code{species_name} from SPECIATE species
records.
}
