% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.cor.R
\name{sp.cor}
\alias{sp.cor}
\alias{sp_profile_cor}
\title{sp_profile correlation}
\usage{
sp_profile_cor(x, min.n = 3, output = c("plot", "report"))
}
\arguments{
\item{x}{A \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}

\item{min.n}{Numeric (default 3), the minimum number of species measurements
needed in a profile for the function to use it in correlation caluclations.
Here, it should be noted that this does not guarantee at the three matched
pairs of measurements needed to calculate a correlation coefficient because
not all profiles contain all species, so there may still be insufficient
overlap on a case-by-case basis.}

\item{output}{Character vector, required function output: \code{'report'} the
calculated correlation matrix; \code{'plot'} a heat map of that correlation
matrix.}
}
\value{
Depending on the \code{output} option, \code{sp_profile_cor} returns
one or more of the following: the correlation matrix, a heat map of the
correlation matrix.
}
\description{
sp_profile functions for studying relationships between
multi-profile (re)SPECIATE data sets

\code{\link{sp_profile_cor}} generates a correlation matrix
for the supplied (re)SPECIATE data sets.
}
