% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.find.R
\name{sp.find}
\alias{sp.find}
\alias{sp_find_profile}
\alias{sp_find_species}
\title{re(SPECIATE) sp_find_ functions}
\usage{
sp_find_profile(..., by = "keywords", partial = TRUE)

sp_find_species(..., by = "species_name", partial = TRUE)
}
\arguments{
\item{...}{character(s), any search term(s) to use when searching
the local (re)SPECIATE archive for relevant records.}

\item{by}{character, the section of the archive to
search, by default \code{'keywords'} for \code{\link{sp_find_profile}} and
\code{'species_names'} for \code{\link{sp_find_species}}.}

\item{partial}{logical, if \code{TRUE} (default)
\code{sp_find_} functions use partial matching.}
}
\value{
\code{sp_find_profile} returns a object of
\code{respeciate.ref} class, a \code{data.frame} of
profile information.
\code{sp_find_species} returns a object of
\code{respeciate.spcs} class, a \code{data.frame} of
species information.
}
\description{
\code{sp_find_} functions are for use with (re)SPECIATE
source data in R...

\code{sp_find_} functions search main data sets
in the (re)SPECIATE archive using supplied search terms.
\code{\link{sp_find_profile}} searches for profile records and
\code{\link{sp_find_species}} searches for species records.
}
\examples{
\dontrun{
profile <- "Ethanol"
pr <- sp_find_profile(profile)
pr

species <- "Ethanol"
sp <- sp_find_species(species)
sp}

}
