% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.info.R
\name{sp.info}
\alias{sp.info}
\alias{sp_info}
\title{re(SPECIATE) information}
\usage{
sp_info()
}
\value{
\code{sp_info} provide brief source information report on the
currently installed (re)SPECIATE data sets.
}
\description{
\code{sp_info} provides (re)SPECIATE
source information...
}
