% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.pad.R
\name{sp.pad}
\alias{sp.pad}
\alias{sp_pad}
\title{(re)SPECIATE profile padding functions}
\usage{
sp_pad(x, pad = "species", drop.nas = TRUE)
}
\arguments{
\item{x}{A \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}

\item{pad}{character, type of meta data padding, current options
\code{'profile'}, \code{species} or \code{weight}.}

\item{drop.nas}{logical, dicard any rows where \code{WEIGHT_PERCENT} is
\code{NA}}
}
\value{
\code{sp_pad} returns \code{x}, with requested additional profile
and species meta-data added as additional \code{data.frame} columns.
See Note.
}
\description{
Functions for padding \code{respeciate} objects.

\code{sp_pad} pads a supplied (re)SPECIATE profile data set
with profile and species meta-data.
}
\note{
Some data handling can remove (re)SPECIATE meta-data,
and \code{sp_pad}s provide a quick rebuild/repair. For example,
\code{\link{sp_dcast}}ing to a (by-species or by-profile) widened
form strips some meta-data, and padding is used as part of the
\code{\link{sp_melt_wide}} and padding is used to re-add this meta-data
when returning the data set to its standard long form.
}
