% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_profile_cluster.R
\name{sp.profile.cluster}
\alias{sp.profile.cluster}
\alias{sp_profile_dist}
\title{sp_profile clustering}
\usage{
sp_profile_dist(x, output = c("plot", "report"))
}
\arguments{
\item{x}{A \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}

\item{output}{Character vector, required function output: \code{'report'} the
calculated correlation matrix; \code{'plot'} a heat map of that distance
matrix.}
}
\value{
Depending on the \code{output} option, \code{sp_profile_dist} returns
one or more of the following: the correlation matrix, a heat map of the
correlation matrix.
}
\description{
sp_profile functions for studying similarities (or
dissimilarities) within multi-profile (re)SPECIATE data sets

\code{\link{sp_profile_dist}} calculates the statistical distance
between re(SPECIATE) profiles, and clusters profiles according to nearness.
}
\note{
Please note: function in development; structure and arguments may be
subject to change.
}
