% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_profile_dcast.R
\name{sp.profile.dcast}
\alias{sp.profile.dcast}
\alias{sp_profile_dcast}
\title{sp_profile reshaping functions}
\usage{
sp_profile_dcast(x, wide = "species")
}
\arguments{
\item{x}{A \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}

\item{wide}{character, when reshaping \code{x} the data type to 'widen',
currently \code{'species'} (default) or \code{'profile'}. See Note.}
}
\value{
\code{sp_profile_dcast} returns the wide form of the supplied
\code{respeciate} profile.
}
\description{
sp_profile functions for reshaping (re)SPECIATE profiles

\code{\link{sp_profile_dcast}} reshapes the supplied
(re)SPECIATE profile(s), converting it from long to short form. See Notes
for further details.
}
\note{
Conventional long-to-wide reshaping of data, or \code{dcast}ing, can
be slow and memory inefficient. So, \code{respeciate} uses the
\code{\link{data.table::dcast}} method. The default method,
applied using \code{wide='species'}, is effectively:

\code{dcast(..., PROFILE_CODE+PROFILE_NAME~SPECIES_NAME, value.var="WEIGHT_PERCENT")}

And, the alternative \code{wide='profile'}:

\code{dcast(..., SPECIES_ID + SPECIES_NAME ~PROFILE_CODE, value.var="WEIGHT_PERCENT")}

Although, \code{respeciate} uses a local version of \code{WEIGHT_PERCENT} called
\code{.value}, so the EPA source information can easily be recovered. See also
\code{\link{sp_profile_rescale}}
}
\references{
Dowle M, Srinivasan A (2023). _data.table: Extension of `data.frame`_.
  R package version 1.14.8, <https://CRAN.R-project.org/package=data.table>.
}
