% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{sp}
\alias{sp}
\alias{sp_profile}
\title{sp_ functions}
\usage{
sp_profile(code)
}
\arguments{
\item{code}{character or numeric, the SPECIATE code
of the required profile (EPA SPECIATE term PROFILE_CODE).}
}
\value{
\code{sp_profile} returns a object of
\code{respeciate} class, a \code{data.frame} containing a
speciate profile.
}
\description{
sp function to get profiles from the R (re)SPECIATE archive

\code{\link{sp_profile}} extracts a
SPECIATE profile from the local (re)SPECIATE archive.
}
\examples{
\dontrun{
x <- sp_profile(c(8833, 8850))
plot(x)}
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
