% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciate.R
\name{spec}
\alias{spec}
\title{Access to the SPECIATE 5.1 US/EPA Tool}
\usage{
spec(code)
}
\arguments{
\item{code}{Character, PROFILE CODE required by EPA/Speciate}
}
\value{
a data.frame with full information for the desired code (PROFILE_CODE)
}
\description{
\code{\link{spec}} Return a speciate data.frame
}
\examples{
\dontrun{
code <- "8855"
x <- spec(code)
}
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
