% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIEUROPE.R
\docType{data}
\name{SPECIEUROPE}
\alias{SPECIEUROPE}
\title{SPECIEUROPE}
\format{
A ( 3  long) 'list' object
\describe{
  \item{source}{The main SPECIEUROPE data set}
  \item{ref}{The source citation, to be used whenever this data is used.}
  \item{website}{The SPECIEUROPE project website link}
}
}
\source{
https://source-apportionment.jrc.ec.europa.eu/
}
\usage{
SPECIEUROPE
}
\description{
The SPECIEUROPE data set is a local version of the European
Commission (EC) Joint Research Centre JRC's repository of particulate matter
(PM) speciation profiles of European air pollutant sources.

Currently using version 3.0 as of 2025-11-19.
}
\references{
Pernigotti, D., Belis, C.A., Spano, L., 2016. SPECIEUROPE: The
European data base for PM source profiles. Atmospheric Pollution Research,
7(2), pp.307-314. DOI: https://doi.org/10.1016/j.apr.2015.10.007
}
\keyword{datasets}
