% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq.R
\name{spq}
\alias{spq}
\alias{spq_gas}
\alias{spq_other}
\alias{spq_pm}
\alias{spq_pm.ae6}
\alias{spq_pm.ae8}
\alias{spq_pm.cr1}
\alias{spq_pm.simplified}
\title{spq_ quick access to common re(SPECIATE) sub-samples}
\usage{
spq_gas()

spq_other()

spq_pm()

spq_pm.ae6()

spq_pm.ae8()

spq_pm.cr1()

spq_pm.simplified()
}
\value{
\code{spq_} functions typically return a \code{respeciate}
\code{data.frame} of the requested profiles.

For example:

\code{sqr_gas} returns all gaseous profiles (\code{PROFILE_TYPE == 'GAS'}).

\code{sqr_pm} returns all particulate matter (PM) profiles not classified
as a special PM type (\code{PROFILE_TYPE == 'PM'}).

The special PM types are subsets of PM profiles intended for special
applications, and these include \code{sqr_pm.ae6} (type \code{PM-AE6}),
\code{sqr_pm.ae8} (type \code{PM-AE8}), \code{sqr_pm.cr1} (type
\code{PM-CR1}), \code{sqr_pm.simplified} (type \code{PM-Simplified})
ans \code{sqr_other} (\code{PROFILE_TYPE == 'OTHER'}).
}
\description{
\code{spq_} functions are quick access wrappers to commonly
requested re(SPECIATE) sub-samples.
}
