% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respeciate-package.R
\docType{package}
\name{respeciate-package}
\alias{respeciate}
\alias{respeciate-package}
\title{respeciate: Speciation profiles for gases and aerosols}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access to the US.EPA Speciate (v5.2) tool, to generate speciation profiles for gases and particles. More details in Simon et al (2010) \doi{10.5094/APR.2010.026}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atmoschem/respeciate}
  \item Report bugs at \url{https://github.com/atmoschem/respeciate/issues}
}

}
\author{
\strong{Maintainer}: Sergio Ibarra-Espinosa \email{sergio.ibarra@usp.br} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})

Authors:
\itemize{
  \item Karl Ropkins \email{k.ropkins@its.leeds.ac.uk} (\href{https://orcid.org/0000-0002-0294-6997}{ORCID})
}

}
\keyword{internal}
