% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respeciate.generics.R
\name{respeciate.generics}
\alias{respeciate.generics}
\alias{print.respeciate}
\alias{plot.respeciate}
\alias{summary.respeciate}
\title{respeciate.generics}
\usage{
\method{print}{respeciate}(x, n = NULL, ...)

\method{plot}{respeciate}(x, n = NULL, id = NULL, order = TRUE, ...)

\method{summary}{respeciate}(object, ...)
}
\arguments{
\item{x}{the \code{respeciate}
object to be printed, plotted, etc.}

\item{n}{when plotting or printing a multi-profile object, the
maximum number of profiles to report. (When plotting, \code{n}
is ignored if \code{id} is also set.)}

\item{...}{any extra arguments, mostly ignored except by
\code{plot} which passes them to \code{\link{barplot}}.}

\item{id}{numeric, indices of profiles to use when
plotting (\code{id=1:6} is equivalent to \code{n=6}).}

\item{order}{logical, order the species in the
profile(s) by relative abundance before plotting.}

\item{object}{like \code{x} but for \code{summary}.}
}
\description{
\code{respeciate} object classes and generic functions.

When supplied a \code{respeciate}
object, \code{\link{print}} manages its appearance.

When supplied a \code{respeciate}
object, \code{\link{plot}} provides a basic plot
output. This uses base function \code{\link{barplot}};
also see note.
}
\note{
\code{respeciate} objects revert to
\code{data.frame}s when not doing anything
package-specific, so you can still
use as previously with \code{lattice} or
\code{ggplot2}, useful if you are pulling multiple
profiles and you exceed the base \code{\link{barplot}}
capacity...
}
