% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciate.0.2.R
\name{respeciate}
\alias{respeciate}
\alias{sp_find_profile}
\alias{sp_profile}
\alias{print.respeciate}
\alias{plot.respeciate}
\alias{print.respeciate.ref}
\title{respeciate}
\usage{
sp_find_profile(term, by = "Keywords")

sp_profile(code)

\method{print}{respeciate}(x, n = 6, ...)

\method{plot}{respeciate}(x, n = NULL, order = TRUE, ..., legend.text = NULL, args.legend = NULL)

\method{print}{respeciate.ref}(x, ...)
}
\arguments{
\item{term}{character, the search term to use
when searching archive.}

\item{by}{character, the section of the archive to
search, by default \code{'keywords'}.}

\item{code}{character or numeric, the speciate code
of the required source profile (EPA SPECIATE term PROFILE_CODE).}

\item{x}{the \code{respeciate} or \code{respeciate.ref}
object to be printed, plotted, etc.}

\item{n}{when plotting or printing a multi-profile, the
number(s) of profile(s) to report.}

\item{...}{any extra arguments, mostly ignored except by
\code{plot} which passes them to \code{\link{barplot}}.}

\item{order}{logical, order the species in the
profile(s) by molecular weight before plotting.}

\item{legend.text, args.legend}{\code{\link{barplot}}
arguments.}
}
\value{
\code{sp_find_profile} returns a object of
\code{respeciate.ref} class, a \code{data.frame} of
profile information.

\code{sp_profile} returns a object of
\code{respeciate} class, a \code{data.frame} containing a
speciate profile.
}
\description{
This is respeciate but with object classes...

\code{\link{sp_find_profile}} finds
the codes of speciate profiles in the local speciate
archive using supplied search terms.

\code{\link{sp_profile}} extracts a
speciate profile from the local speciate archive.

When supplied a \code{respeciate}
object, \code{\link{print}} manages its appearance.

When supplied a \code{respeciate}
object, \code{\link{plot}} provides a basic plot
output. This uses base function \code{\link{barplot}};
also see note.

When supplied a \code{respeciate.ref}
object, \code{\link{print}} manages its appearance.
}
\note{
\code{respeciate} objects revert to
\code{data.frame}s when not doing anything
package-specific, so you can still
use as previously with \code{lattice} or
\code{ggplot2}, useful it you are pulling multiple
profiles and you exceed the base \code{\link{barplot}}
capacity...
}
\examples{
\dontrun{
profile <- "Ethanol"
dt <- sp_find_profile(profile)
}
\dontrun{
x <- sp_profile(c(8833, 8850))
plot(x)
}
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
