% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.match.R
\name{rsp.match}
\alias{rsp.match}
\alias{rsp_match_profile}
\title{Find nearest matches from reference set of profiles}
\usage{
rsp_match_profile(
  rsp,
  ref,
  matches = 10,
  rescale = 5,
  min.n = NULL,
  method = "sid * srd",
  self.test = FALSE,
  ...,
  output = "summary"
)
}
\arguments{
\item{rsp}{A \code{respeciate} object or similar \code{data.frame} containing
a species profile to be compared with profiles in \code{ref}. If \code{rsp}
contains more than one profile, these are averaged (using
\code{\link{rsp_average_profile}}), and the average compared.}

\item{ref}{A \code{respeciate} object, a \code{data.frame} containing a
multiple species profiles, to be used as reference library when identifying
nearest matches for \code{rsp}.}

\item{matches}{Numeric (default 10), the maximum number of profile matches to
report.}

\item{rescale}{Numeric (default 5), the data scaling method to apply before
comparing \code{rsp} and profiles in \code{ref}: options 0 to 5 handled by
\code{\link{rsp_rescale}}.}

\item{min.n}{Numeric (or \code{NULL}), the minimum number of paired species
measurements required for a match to be assessed. The larger \code{min.n},
the greater the required \code{rsp} and \code{ref} profile overlap, so the
better the matching confidence for paired cases but also the more likely
that a sparse but relevant \code{ref} profile may be missing. The default
option, \code{NULL}, is 65\% of the number of species in \code{rsp} or 6
if larger.}

\item{method}{Character (default 'sid * srd'), the ranking metric used to
rank profile matches. The function calculates several matching metrics:
'pd', the Pearson's Distance (1 - Pearson's correlation coefficient),
'srd', like pd but using the Spearman Ranked data correlation coefficient,
and 'sid', the Standardized Identity Distance (See References). All the
metrics tend to zero for better matches, and the \code{method} can be
any character string that can be evaluated from any of these, e.g.,
\code{'pd'}, \code{'srd'}, \code{'sid'}, and combinations thereof.}

\item{...}{Additional arguments, typically ignore but sometimes used for
function development. Currently, testing \code{rm.reps} (logical) option to
remove what appear to be replicate profile matches from the result set. This
is based on the assumption that identical 'pd' and 'sid' scores identical
identical \code{ref} profiles (or identical overlaps with \code{rsp}) but is
not validated, so handle with care...}

\item{output}{Character, output options, including: \code{'summary'} (the
default) a \code{data.frame} of the requested best \code{matches}, ranked
according to the \code{method} used; \code{'plot'} the associated output from
\code{\link{rsp_plot_match}}.}

\item{slef.test}{Logical (default FALSE). The match process self-tests by adding
\code{rsp} to \code{ref}, which should generate an ideal (nearness = 0) score.
Setting \code{self.test} to \code{TRUE} retains this as an extra record.}
}
\value{
By default \code{rsp_match_profile} returns a fit report summary: a
\code{data.frame} of up to \code{matches} fit reports for the nearest
matches to profiles from the reference profile data set, \code{ref}. (See
also \code{output} above for other options). If several options are requested,
earlier options are report (e.g. using \code{print} or \code{plot}) and only
the final option is returned.
}
\description{
\code{rsp_match_profile} compares a supplied respeciate
profile (or similar data set) and a reference set of supplied profiles
and attempts to identify nearest matches on the
basis of similarity.
}
\references{
Distance metrics are based on recommendations by Belis et al (2015)
and as implemented in Mooibroek et al (2022):

Belis, C.A., Pernigotti, D., Karagulian, F., Pirovano, G., Larsen, B.R.,
Gerboles, M., Hopke, P.K., 2015. A new methodology to assess the performance
and uncertainty of source apportionment models in intercomparison
exercises. Atmospheric Environment, 119, 35–44.
https://doi.org/10.1016/j.atmosenv.2015.08.002.

Mooibroek, D., Sofowote, U.M. and Hopke, P.K., 2022. Source apportionment of
ambient PM10 collected at three sites in an urban-industrial area with
multi-time resolution factor analyses. Science of The Total Environment,
850, p.157981. http://dx.doi.org/10.1016/j.scitotenv.2022.157981.
}
\seealso{
\code{\link{rsp_plot_match}}
}
