% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsp.R
\name{rsp}
\alias{rsp}
\alias{rsp_profile}
\title{rsp_profile}
\usage{
rsp(..., include.refs = FALSE, source = "us")

rsp_profile(...)
}
\arguments{
\item{...}{The function assumes all inputs (except \code{include.refs}
and \code{source}) are profile identifiers: namely, \code{PROFILE_CODE}
and \code{Species.id} in SPECIATE and SPECIEUROPE, respectively, or
potential sources of profile information and requests these
form the local respeciate archives. Typically, simple
objects like character and numeric vectors, as assumed to profile codes and
composite data-types like \code{respeciate} or \code{data.frame} objects
are assumed to contain a named \code{PROFILE_CODE} column. All potential
profile codes are requested and unrecognized codes (and code duplicates)
are ignored.}

\item{include.refs}{logical, if profile reference information should be
included when extracting the requested profile(s) from the archive, default
\code{FALSE}.}

\item{source}{character, the local archive to request a profile from:
\code{'us'} US EPA SPECIATE, or \code{'eu'} EU JRC SPECIEUROPE.}
}
\value{
\code{rsp_profile} or the short-hand \code{rsp} return an object of
\code{respeciate} class, a \code{data.frame} containing one or more profile
from the local respeciate archive.
}
\description{
Getting profile(s) from the R respeciate archive
}
\note{
The option \code{include.refs} adds profile source reference
information to the returned \code{respeciate} data set. The default option
is to not include these because some SPECIATE profiles have several
associated references and including these replicates records, once per
reference.
\code{respeciate} code is written to handle this but if you are developing
own methods or code and include references in any profile build you may be
biasing some analyses in favor of those multiple-reference profile unless
you check and account such cases.
}
\examples{
\dontrun{
x <- rsp_profile(8833, 8850)
plot(x)}
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
\seealso{
\code{\link{SPECIATE}}
}
