% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.001.R
\name{sp.001}
\alias{sp.001}
\alias{sp_find_profile}
\alias{sp_find_species}
\alias{sp_profile}
\title{sp_ functions}
\usage{
sp_find_profile(..., by = "keywords", partial = TRUE)

sp_find_species(..., by = "species_name", partial = TRUE)

sp_profile(code)
}
\arguments{
\item{...}{for \code{sp_find} functions, character(s), any
search term(s) to use when searching the local (re)SPECIATE archive for
relevant records.}

\item{by}{character, the section of the archive to
search, by default \code{'keywords'} for \code{\link{sp_find_profile}} and
\code{'species_names'} for \code{\link{sp_find_species}}.}

\item{partial}{logical, if \code{TRUE} (default)
\code{sp_find} functions use partial matching.}

\item{code}{character or numeric, the SPECIATE code
of the required profile (EPA SPECIATE term PROFILE_CODE).}
}
\value{
\code{sp_find_profile} returns a object of
\code{respeciate.ref} class, a \code{data.frame} of
profile information.

\code{sp_find_species} returns a object of
\code{respeciate.spcs} class, a \code{data.frame} of
species information.

\code{sp_profile} returns a object of
\code{respeciate} class, a \code{data.frame} containing a
speciate profile.
}
\description{
sp functions for use with (re)speciate data in R...

\code{sp_find} functions search main data sets
in the (re)SPECIATE archive using supplied search terms.
\code{\link{sp_find_profile}} searches for profile records and
\code{\link{sp_find_species}} searches for species records.

\code{\link{sp_profile}} extracts a
SPECIATE profile from the local (re)SPECIATE archive.
}
\examples{
\dontrun{
profile <- "Ethanol"
dt <- sp_find_profile(profile)
dt}

\dontrun{
species <- "Ethanol"
sp <- sp_find_species(species)
sp}

\dontrun{
x <- sp_profile(c(8833, 8850))
plot(x)}
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
