% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.cor.R
\name{sp.cor}
\alias{sp.cor}
\alias{sp_species_cor}
\title{(re)SPECIATE Species Correlations}
\usage{
sp_species_cor(
  x,
  min.n = 3,
  cols = c("#80FFFF", "#FFFFFF", "#FF80FF"),
  na.col = "#CFCFCF",
  output = c("plot", "report")
)
}
\arguments{
\item{x}{\code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}

\item{min.n}{\code{numeric} (default 3), the minimum number of species measurements
needed in a profile for the function to use it in correlation calculations.
Here, it should be noted that this does not guarantee the three matched
pairs of measurements needed to calculate a correlation coefficient because
not all profiles contain all species, so there may still be insufficient
overlap on a case-by-case basis.}

\item{cols}{a series of \code{numeric}, \code{character} or other class values
that can be translated into a color gradient, used to color valid cases when
generating plots and color keys, default \code{c("#80FFFF", "#FFFFFF", "#FF80FF")}
equivalent to \code{\link{cm.colors}} output.}

\item{na.col}{\code{numeric}, \code{character} or other class that can be
translated into a single color, used to color \code{NA}s when generating
plots and color keys, default grey \code{"#CFCFCF"}.}

\item{output}{\code{character} vector, required function output, one or more
of : \code{'report'} the calculated correlation matrix; \code{'plot'} a heat
map of that correlation matrix.}
}
\value{
Depending on the \code{output} option, \code{sp_species_cor} returns
one or more of the following: the correlation matrix, a heat map of the
correlation matrix.
}
\description{
sp_species functions for studying relationships between
species in multi-profile (re)SPECIATE data sets.

\code{\link{sp_species_cor}} generates a by-species correlation
matrix of the supplied (re)SPECIATE data sets.
}
