% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.info.R
\name{sp.info}
\alias{sp.info}
\alias{sp_info}
\alias{sp_profile_info}
\alias{sp_species_info}
\alias{sp_find_profile}
\alias{sp_find_species}
\title{re(SPECIATE) information}
\usage{
sp_info()

sp_profile_info(..., by = "keywords", partial = TRUE)

sp_find_profile(...)

sp_species_info(..., by = "species_name", partial = TRUE)

sp_find_species(...)
}
\arguments{
\item{...}{character(s), any search term(s) to use when searching
the local (re)SPECIATE archive for relevant records using
\code{sp_profile_info} or \code{sp_species_info}.}

\item{by}{character, the section of the archive to
search, by default \code{'keywords'} for \code{sp_profile_info} and
\code{'species_names'} for \code{sp_species_info}.}

\item{partial}{logical, if \code{TRUE} (default)
\code{sp_profile_info} or \code{sp_profile_info} use partial matching.}
}
\value{
\code{sp_info} provides a brief version information report on the
currently installed (re)SPECIATE archive.

\code{sp_profile_info} returns a \code{data.frame} of
profile information, as a \code{respeciate} object.
\code{sp_species_info} returns a \code{data.frame} of
species information as a \code{respeciate} object.
}
\description{
Functions that provide (re)SPECIATE
source information.
\code{sp_info} generates a brief version report for the currently installed
(re)SPECIATE data sets.
\code{sp_profile_info} searches the currently installed (re)SPECIATE
data sets for profile records.
\code{sp_species_info} searches the currently installed (re)SPECIATE
data sets for species records.
}
\examples{
\dontrun{
profile <- "Ethanol"
pr <- sp_find_profile(profile)
pr

species <- "Ethanol"
sp <- sp_find_species(species)
sp}

}
