% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.match.R
\name{sp.match}
\alias{sp.match}
\alias{sp_match_profile}
\title{Find nearest matches from reference set of profiles}
\usage{
sp_match_profile(x, ref, n = 10, rescale = 4)
}
\arguments{
\item{x}{A \code{respeciate} object, a \code{data.frame} containing a
species profile to be compared with profiles in \code{ref}.}

\item{ref}{A \code{respeciate} object, a \code{data.frame} containing a
multiple species profiles, to be used as reference library when identifying
matches with \code{x}.}

\item{n}{Numeric (default 10), the maximum number of profile matches to
report.}

\item{rescale}{Numeric (default 4), the data scaling method to apply before
comparing \code{x} and profiles in \code{ref}: options 0 to 4 handled by
\code{\link{sp_rescale_profile}}.}
}
\value{
\code{sp_match_profile} returns a fit report: a \code{data.frame} of
up to \code{n} fit reports for the nearest matches to \code{x} from the
reference profile data set, \code{ref}.
}
\description{
sp_profile functions identify nearest matches in a reference
set of (re)SPECIATE profiles for a supplied species profile.

\code{sp_match_profile} attempts to identify the nearest
matches to a supplied profile from a reference set of (re)SPECIATE profiles
on the basis of correlation coefficient.
}
