% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.plot.R
\name{sp.plot}
\alias{sp.plot}
\alias{sp_plot_profile}
\alias{sp_plot_species}
\title{plotting (re)SPECIATE profiles}
\usage{
sp_plot_profile(
  x,
  id,
  multi.profile = "group",
  order = TRUE,
  log = FALSE,
  ...,
  silent = FALSE
)

sp_plot_species(
  x,
  id,
  multi.species = "group",
  order = FALSE,
  log = FALSE,
  ...,
  silent = FALSE
)
}
\arguments{
\item{x}{A \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}

\item{id}{numeric, indices of profiles to use when
plotting (e.g. \code{id=1:6} plots first 6 profiles).}

\item{multi.profile}{character, how plot should handle
multiple profiles, e.g. 'group' or 'panel' (default
group).}

\item{order}{logical, order the species in the
profile(s) by relative abundance before plotting.}

\item{log}{logical, log y scale when plotting.}

\item{...}{any additional arguments, typically passed on the lattice
plotting functions.}

\item{silent}{logical, hide warnings when generating plots (default
\code{FALSE})}

\item{multi.species, }{like \code{multi.profile} but for species.}
}
\value{
\code{sp_plot} graph, plot, etc usually as a trellis object.
}
\description{
Functions for plotting \code{respeciate} objects.

\code{sp_plot} functions generate plots for supplied
(re)SPECIATE profile data sets.
}
\note{
These functions are currently in development, so may change.
}
