% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.pls.R
\name{sp.pls}
\alias{sp.pls}
\alias{sp_pls_profile}
\alias{pls_report}
\alias{pls_refit_species}
\alias{pls_add_parent}
\title{(re)SPECIATE profile Positive Least Squares}
\usage{
sp_pls_profile(x, ref, power = 1, ...)

pls_report(pls)

pls_refit_species(pls, name, power = 1, ...)

pls_add_parent(pls, parent, power = 1, ...)
}
\arguments{
\item{x}{A \code{respeciate} object, a \code{data.frame} of
profiles in standard long form, intended for PLS modelling.}

\item{ref}{A \code{respeciate} object, a \code{data.frame} of
profiles also in standard long form, intended to be used as sample
of references profiles to fit \code{x} as a linear combination
of. See also Notes.}

\item{power}{A numeric, an additional factor to be added to
weightings when fitting the PLS model. This is applied in the form
\code{weight^power}, and increasing this, increases the relative
weighting of the more heavily weighted measurements. Values in the
range \code{1 - 2.5} are sometimes helpful.}

\item{...}{additional arguments, currently ignored.}

\item{pls}{A \code{sp_pls_profile} output, only used by \code{pls_}
functions.}

\item{name}{Character class, for \code{pls_refit_species} only, the
name of the species to refit. All other species in the reference
profiles are held 'as is' and the \code{name}d species is refit based on
previous PLS estimated source contribution time-series, then these are
in turn refit using the new profiles generated by re-distributing the
\code{name}d species.}

\item{parent}{(for \code{pls_add_parent} only) a data.frame of measurements
of an additional species. This is treated as the sample parent (or a proxy
for this), e.g. the total particulate matter (PM) measurement for
(percentage) PM profiles. It is added to \code{x}, dummy entries are added
to \code{ref} with a start value of 100 (percent), and the model is refitted.}
}
\value{
\code{sp_pls_profile} returns a list of nls models, one per
profile/measurement set in \code{x}. The \code{pls_} functions work with
these outputs. \code{pls_report} generates a \code{data.frame} of
model outputs, and is used of several of the other \code{pls_}
functions. \code{pls_refit_species} and \code{pls_add_daughter}
return the supplied \code{sp_pls_profile} output, updated on the
basis of the \code{pls_} function call.
}
\description{
Functions for Positive Least Squares (PSL) fitting of
(re)SPECIATE profiles

\code{sp_pls_profile} builds PSL models for supplied profile(s) using
the \code{\link{nls}} function, the 'port' algorithm and a lower
limit of zero for all model outputs to enforce the positive fits. The
modeled profiles are typically from an external source, e.g. a
measurement campaign, and the models are fit using a set of reference
profiles, here typically from (re)SPECIATE, to provide a measure of
source apportionment based on the assumption that the profiles in the
reference set are representative of the mix that make up the modeled
sample. The \code{pls_} functions work with \code{sp_pls_profile}
outputs, and are intended to be used when refining and analyzing
these PLS models.
}
\note{
This implementation of PLS applies the following modeling constaints:

1. The generate a model of \code{x} that is positively constained linear
product of the profiles in \code{ref}, so outputs can only be
zero or more.  Although the model is generated using \code{\link{nls}},
which is a Nonlinear Least Squares (NLS) model, the fitting term applied
in this case is linear.

2. The number of species in \code{x} must be more that the number of
profiles in \code{ref}, to reduce the likelihood of over-fitting.
}
