% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.pls.R
\name{sp.pls}
\alias{sp.pls}
\alias{sp_pls_profile}
\alias{pls_report}
\alias{pls_fit_species}
\alias{pls_refit_species}
\alias{pls_rebuild}
\alias{pls_plot}
\alias{pls_plot_species}
\alias{pls_plot_profile}
\title{(re)SPECIATE profile Positive Least Squares}
\usage{
sp_pls_profile(x, ref, power = 1, ...)

pls_report(pls)

pls_fit_species(pls, species, power = 1, refit.profile = TRUE, ...)

pls_refit_species(pls, species, power = 1, refit.profile = TRUE, ...)

pls_rebuild(pls, species, power = 1, refit.profile = TRUE, ...)

pls_plot(pls, n = 1, type = 1, ...)

pls_plot_species(pls, n, type = 1, ...)

pls_plot_profile(pls, n, log = FALSE, ...)
}
\arguments{
\item{x}{A \code{respeciate} object, a \code{data.frame} of
profiles in standard long form, intended for PLS modelling.}

\item{ref}{A \code{respeciate} object, a \code{data.frame} of
profiles also in standard long form, used as the set of candidate
source profiles when fitting \code{x}.}

\item{power}{A numeric, an additional factor to be added to
weightings when fitting the PLS model. This is applied in the form
\code{weight^power}, and increasing this, increases the relative
weighting of the more heavily weighted measurements. Values in the
range \code{1 - 2.5} are sometimes helpful.}

\item{...}{additional arguments, typically ignored or passed on to
\code{\link{nls}}.}

\item{pls}{A \code{sp_pls_profile} output, only used by \code{pls_}
functions.}

\item{species}{for \code{pls_fit_species}, a data.frame of
measurements of an additional species to be fitted to an existing
PLS model, or for \code{pls_refit_species} a character vector of the
names of species already included in the model to be refit. Both are
multiple-\code{species} wrappers for \code{pls_rebuild}, a general-purpose
PLS fitter than only handles single \code{species}.}

\item{refit.profile}{(for \code{pls_fit_species}, \code{pls_refit_species}
and \code{pls_rebuild}) logical. When fitting a new \code{species} (or
refitted an existing \code{species}), all other species in the reference
profiles are held 'as is' and added \code{species} is fit to the source
contribution time-series of the previous PLS model. By default, the full PLS
model is then refit using the revised \code{ref} source profile to generate
a PLS model based on the revised source profiles (i.e., ref + new species
or ref + refit species). However, this second step can be omitted using
\code{refit.profile=FALSE} if you want to use the supplied \code{species}
as an indicator rather than a standard member of the apportionment model.}

\item{n}{(for \code{pls_plot}s only) numeric or character
identifying the species or profile to plot. If numeric, these are treated
as indices of the species or profile, respectively, in the PLS model; if
character, species is treated as the name of species and profile is treated
as the profile code. Both can be concatenated to produce multiple plots and
the special case \code{n = -1} is a short cut to all species or profiles,
respectively.}

\item{type}{(for \code{pls_plot}s only) numeric, the plot type if
multiple options are available.}

\item{log}{(for \code{pls_plot_profile} only) logical, if \code{TRUE} this
applies 'log' scaling to the primary Y axes of the plot.}
}
\value{
\code{sp_pls_profile} returns a list of nls models, one per
profile/measurement set in \code{x}. The \code{pls_} functions work with
these outputs. \code{pls_report} generates a \code{data.frame} of
model outputs, and is used of several of the other \code{pls_}
functions. \code{pls_fit_species}, \code{pls_refit_species} and
\code{pls_fit_parent} return the supplied \code{sp_pls_profile} output,
updated on the basis of the \code{pls_} function action.
\code{pls_plot}s produce various plots commonly used in source
apportionment studies.
}
\description{
Functions for Positive Least Squares (PSL) fitting of
(re)SPECIATE profiles

\code{sp_pls_profile} builds PSL models for supplied profile(s) using
the \code{\link{nls}} function, the 'port' algorithm and a lower
limit of zero for all model outputs to enforce the positive fits. The
modeled profiles are typically from an external source, e.g. a
measurement campaign, and are fit as a linear additive series of reference
profiles, here typically from (re)SPECIATE, to provide a measure of
source apportionment based on the assumption that the profiles in the
reference set are representative of the mix that make up the modeled
sample. The \code{pls_} functions work with \code{sp_pls_profile}
outputs, and are intended to be used when refining and analyzing
these PLS models.
}
\note{
This implementation of PLS applies the following modeling constraints:

1. It generates a model of \code{x} that is positively constrained linear
product of the profiles in \code{ref}, so outputs can only be
zero or more.  Although the model is generated using \code{\link{nls}},
which is a Nonlinear Least Squares (NLS) model, the fitting term applied
in this case is linear.

2. The number of species in \code{x} must be more that the number of
profiles in \code{ref} to reduce the likelihood of over-fitting.
}
