% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{sp}
\alias{sp}
\alias{sp_profile}
\title{sp_ functions}
\usage{
sp_profile(code, include.refs = FALSE)
}
\arguments{
\item{code}{character, numeric or data.frame, the SPECIATE code
of the required profile (EPA SPECIATE identifier PROFILE_CODE). This is
typically one or concatenated character or numeric entries, but can also
be a \code{respeciate} object or similar \code{data.frame} containing
the \code{code}s as a named \code{PROFILE_NAME} column.}

\item{include.refs}{logical, include profile reference information when
getting the requested profile(s) from the archive, default \code{FALSE}.}
}
\value{
\code{sp_profile} returns a object of
\code{respeciate} class, a \code{data.frame} containing a
(re)SPECIATE profile.
}
\description{
sp function to get profiles from the R (re)SPECIATE archive

\code{\link{sp_profile}} extracts a
SPECIATE profile from the local (re)SPECIATE archive.
}
\note{
The option \code{include.refs} adds profile source reference
information to the returned \code{respeciate} data set. The default option
is to not include these because some profiles have several associated
references and including these replicates records, once per reference.
\code{respeciate} code is written to handle this but if you are developing
own methods or code and include references in any profile build you may be
biasing some analyses in favor of those multiple-reference profile unless
you check and account such cases.
}
\examples{
\dontrun{
x <- sp_profile(c(8833, 8850))
plot(x)}
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
