% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIATE.R
\docType{data}
\name{SPECIATE}
\alias{SPECIATE}
\title{SPECIATE}
\format{
A ( 13  long) 'list' object
\describe{
  \item{PROFILES}{The main \code{data.frame} of profile-specific meta-data,
  with one row per profile, key term \code{PROFILE_CODE}.}
  \item{SPECIES}{The main \code{data.frame} of individual record meta-data,
  with one row per species in each profile, key terms \code{PROFILE_CODE}
  and \code{SPECIES_ID} linking \code{PROFILES} and
  \code{SPECIES_PROPERTIES}.}
  \item{SPECIES_PROPERTIES}{The main \code{data.frame} of species-specific
  meta-data, with one row per species, key term \code{SPECIES_ID}.}
  \item{PROFILE_REFERENCE}{The \code{data.frame} linking profile and
  reference meta-data, one row per references per profile, key terms
  \code{PROFILE_CODE} and \code{REF_Code}.}
  \item{REFERENCES}{The main \code{data.frame} of references for profile
  source meta-data, one row per reference, key term \code{REF_Code}.}
  \item{And others}{Currently not documented.}
}
}
\source{
https://www.epa.gov/air-emissions-modeling/speciate
}
\usage{
SPECIATE
}
\description{
the SPECIATE data set is a local version of the EPA's
SPECIATE repository of organic gas and particulate matter (PM) speciation
profiles of air pollution sources.
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
\keyword{datasets}
