% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spx.R
\name{spx}
\alias{spx}
\alias{spx_copy}
\alias{spx_}
\alias{spx_n_alkane}
\alias{spx_btex}
\title{spx_ functions for grouping and subsetting}
\usage{
spx_copy(x, ref = NULL, by = "species_id")

spx_n_alkane(x)

spx_btex(x)
}
\arguments{
\item{x}{a \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}

\item{ref}{(\code{spx_copy} only) a second \code{respeciate} object, to
be used as reference when testing \code{x}.}

\item{by}{(\code{spx_copy} only) character, the name of the column
in \code{ref} to copy when testing \code{x}.}
}
\value{
\code{spx_copy} outputs can be modified but, by default, it
identifies all species in the supplied reference data set.

\code{spx_n_alkane} identifies C1 to C40 n-alkanes.

\code{spx_btex} identifies the BTEX group of aromatic hydrocarbons
(benzene, toluene, ethyl benzene, and M-, O- and P-xylene).
}
\description{
\code{spx_} functions generate a vector of assignment
terms and can be used to subset or condition a supplied re(SPECIATE)
\code{data.frame}.

Most commonly, the \code{spx_} functions accept a single input, a
re(SPECIATE) \code{data.frame} and return a logical vector of
length \code{nrow(x)}, identifying species of interest as
\code{TRUE}. So, for example, they can be used when
\code{\link{subset}}ting in the form:

\code{subset(x, spx_n_alkane(x))}

... to extract just n-alkane records from a \code{respeciate} object
\code{x}.

However, some accept additional arguments. For example, \code{spx_copy}
also accepts a reference data set, \code{ref}, and a column identifier,
\code{by}, and tests \code{x$by \%in\% unique(ref$by)}.
}
