% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spq.R, R/spx.R
\name{spx_n_alkane}
\alias{spx_n_alkane}
\alias{spx_btex}
\alias{spx}
\title{spx_ functions for grouping and subsetting}
\usage{
spx_n_alkane(x)

spx_btex(x)

spx_n_alkane(x)

spx_btex(x)
}
\arguments{
\item{x}{a \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}
}
\value{
\code{spx_n_alkane} identifies C1 to C40 n-alkanes.

\code{spx_btex} identifies the BTEX group of aromatic hydrocarbons
(benzene, toulene, ethylbenzene, and M-, O- and P-xylene).
}
\description{
\code{spx_} functions generate a vector of assignment
terms and can be used to subset or condition a supplied
}
