% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spx_.R
\name{spx}
\alias{spx}
\alias{spx_n_alkane}
\alias{spx_btex}
\title{spx_ functions for grouping and subsetting}
\usage{
spx_n_alkane(x)

spx_btex(x)
}
\arguments{
\item{x}{a \code{respeciate} object, a \code{data.frame} of re(SPECIATE)
profiles.}
}
\value{
\code{spx_n_alkane} identifies C1 to C40 n-alkanes.

\code{spx_btex} identifies the BTEX group of aromatic hydrocarbons
(benzene, toulene, ethylbenzene, and M-, O- and P-xylene).
}
\description{
\code{spx_} functions generate a vector of assignment
terms and can be used to subset or condition a supplied
\code{respeciate} object. Most commonly, the \code{spx_} functions
return a logical vector of length = nrow(x), identifying
species of interest as \code{TRUE}. For example:

\code{subset(x, spx_n_alkane(x))}

... to extract just n-alkane records from \code{respeciate} object
\code{x}.
}
