% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{sysdata}
\alias{sysdata}
\title{sysdata}
\format{
A (5  long) 'list' object
\describe{
  \item{PROFILES}{[to document]}
  \item{SPECIES}{[to document]}
  \item{SPECIES_PROPERTIES}{[to document]}
  \item{PROFILE_REFERENCE}{[to document]}
  \item{REFERENCES}{[to document]}
}
}
\source{
[to doc]
}
\usage{
sysdata
}
\description{
splat_packed data [to be described]
}
\references{
Simon, H., Beck, L., Bhave, P.V., Divita, F., Hsu, Y., Luecken, D.,
Mobley, J.D., Pouliot, G.A., Reff, A., Sarwar, G. and Strum, M., 2010.
The development and uses of EPA SPECIATE database.
Atmospheric Pollution Research, 1(4), pp.196-206.
}
\keyword{datasets}
