% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\name{sysdata}
\alias{sysdata}
\title{sysdata}
\format{
A ( 5  long) 'list' object
\describe{
  \item{PROFILES}{[to document]}
  \item{SPECIES}{[to document]}
  \item{SPECIES_PROPERTIES}{[to document]}
  \item{PROFILE_REFERENCE}{[to document]}
  \item{REFERENCES}{[to document]}
}
}
\description{
splat_packed data [to be described]
}
